requirements.txt
# --- Core utilities (lightweight, safe to install early) ---
numpy==1.26.4
pillow==10.4.0
python-dotenv==1.0.1

# --- Dev & testing ---
pytest==8.3.3
pytest-cov==5.0.0

# --- Tracking & data versioning (to be installed later) ---


# --- API for inference (we'll use in M2) ---
fastapi==0.115.4
uvicorn[standard]==0.32.0

# --- Deep Learning framework (chosen) ---
torch==2.10.0+cpu
torchvision==0.25.0+cpu

# --- Experiment tracking ---
mlflow==2.16.0

# --- Training utilities ---
scikit-learn==1.5.2
matplotlib==3.9.2
pandas==2.2.2
seaborn==0.13.2
pyyaml==6.0.2